Added try catch block to the api call to Cloudflare
